fix txns count

+++ src/main/java/Services/TransactionService.java
@@ -39,7 +39,7 @@ public class TransactionService {
             int offset = (page - 1) * count;
 
             List<NewTxn> txns = getTransactions(count, offset);
-            int totalTxnCount = cacheManager.getTotalTransactionCount();
+            long totalTxnCount = cacheManager.getTotalTransactionCount();
 
             for (NewTxn txn : txns) {
                 JSONObject object = populateTxnsResponse(txn);
